refactor: movestrategy 분리
